fixing GPS and BAR initial data

+++ Simulation/Simulation_ekf/types.h
@@ -31,9 +31,9 @@ typedef struct {
 
 // GPS
 typedef struct {
-    float lat = 0.0f;
-    float lon = 0.0f;
-    float alt = 0.0f;
+    double lat = 0.0f;
+    double lon = 0.0f;
+    double alt = 0.0f;
     unsigned long t = 0.0f;
     float dt = 0.0f;
 } gps_t;
